Example script correction; simplified function.
The example script WsusTask.ps1 had a bug where the logs printed didn't
contain the intended information, like wrong variable names being used
for the logging and the fact that exceptions weren't being logged. Also,
the way that the Approve-WsusUpdatesForGroup function was obtaining
update groups was kind of complicated; this was changed to use simpler
cmdlets.